Switch to cudaFreeAsync() and cudaMallocAsync()
cudaFree() can implicitly synchronize, and the CUDA documentation
recommends using cudaFreeAsync() for performance.

To do this, also switch cudaMalloc() to cudaMallocAsync(), to ensure
that the allocations are ordered after the frees (and do not errantly
give out-of-memory errors).

It is safe to not synchronize after calling cudaMallocAsync(), as
all operations directly using GPU memory must run on the GPU, and
those operations are all enqueued in the same stream, ensuring that
the allocation is complete before it's used.